refactor: 답변 태그 부분 파일로 분리
